+++ example/lib/examples/button_builder_example/example.dart
@@ -65,7 +65,7 @@ class _ButtonBuilderExampleState extends State<ButtonBuilderExample> {
             padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
             child: Text(
               'Checkbox buttons',
-              style: theme.textTheme.headline5,
+              style: theme.textTheme.headlineSmall,
             ),
           ),
           GroupButton(
@@ -96,7 +96,7 @@ class _ButtonBuilderExampleState extends State<ButtonBuilderExample> {
             padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
             child: Text(
               'Radio buttons',
-              style: theme.textTheme.headline5,
+              style: theme.textTheme.headlineSmall,
             ),
           ),
           GroupButton(
@@ -114,6 +114,24 @@ class _ButtonBuilderExampleState extends State<ButtonBuilderExample> {
                 },
               );
             },
+            onSelected: (val, i, selected) =>
+                debugPrint('Button: $val index: $i $selected'),
+          ),
+          SizedBox(height: 16),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+            child: Text(
+              'Text builder example',
+              style: theme.textTheme.headlineSmall,
+            ),
+          ),
+          GroupButton(
+            controller: _radioController,
+            isRadio: true,
+            options: GroupButtonOptions(groupingType: GroupingType.wrap),
+            buttons: _radioButtons,
+            buttonIndexedTextBuilder: (selected, index, context) =>
+                _radioButtons[index] + ' [built]',
             onSelected: (val, i, selected) =>
                 debugPrint('Button: $val index: $i $selected'),
           ),