+++ lib/src/group_button_body.dart
@@ -38,6 +38,8 @@ class GroupButtonBody<T> extends StatefulWidget {
     this.elevation,
     this.buttonIndexedBuilder,
     this.buttonBuilder,
+    this.buttonIndexedTextBuilder,
+    this.buttonTextBuilder,
   }) : super(key: key);
 
   final List<T> buttons;
@@ -72,6 +74,8 @@ class GroupButtonBody<T> extends StatefulWidget {
   final GroupButtonController? controller;
   final GroupButtonIndexedBuilder? buttonIndexedBuilder;
   final GroupButtonValueBuilder<T>? buttonBuilder;
+  final GroupButtonIndexedTextBuilder? buttonIndexedTextBuilder;
+  final GroupbuttonTextBuilder<T>? buttonTextBuilder;
 
   @override
   State<GroupButtonBody<T>> createState() => _GroupButtonBodyState<T>();
@@ -162,7 +166,11 @@ class _GroupButtonBodyState<T> extends State<GroupButtonBody<T>> {
         );
       } else {
         button = GroupButtonItem(
-          text: widget.buttons[i].toString(),
+          text: widget.buttonIndexedTextBuilder
+                  ?.call(_isSelected(i), i, context) ??
+              widget.buttonTextBuilder
+                  ?.call(_isSelected(i), widget.buttons[i], context) ??
+              widget.buttons[i].toString(),
           onPressed: _controller.disabledIndexes.contains(i)
               ? () => _controller.onDisablePressed?.call(i)
               : () {