+++ lib/src/utils/models/types.dart
@@ -13,3 +13,17 @@ typedef GroupButtonValueBuilder<T> = Widget Function(
   T value,
   BuildContext context,
 );
+
+/// Custom builder method to create custom buttons by index
+typedef GroupButtonIndexedTextBuilder = String Function(
+  bool selected,
+  int index,
+  BuildContext context,
+);
+
+/// Custom builder method to create custom buttons by value
+typedef GroupbuttonTextBuilder<T> = String Function(
+  bool selected,
+  T value,
+  BuildContext context,
+);